+++ packages/app/components/Proposals/CurrentStandings.tsx
@@ -1,13 +1,9 @@
 import Divider from 'components/CommonComponents/Divider';
 import ProgressBar from 'components/ProgressBar';
-import { BigNumber } from 'ethers';
 import { Proposal } from 'interfaces/interfaces';
 import { DateTime } from 'luxon';
 import { useEffect, useState } from 'react';
-import {
-  bigNumberToNumber,
-  formatAndRoundBigNumber,
-} from '@popcorn/utils/src/formatBigNumber'
+import { bigNumberToNumber, formatAndRoundBigNumber } from 'utils/formatBigNumber';
 
 const getTimeLeft = (stageDeadline: Date): string => {
   const date1 = DateTime.fromISO(new Date().toISOString());
