+++ packages/app/pages/beneficiaries/[id].tsx
@@ -4,6 +4,7 @@ import { useContext, useEffect, useState } from 'react';
 import { getIpfsHashFromBytes32 } from '@popcorn/utils/ipfsHashManipulation';
 import BeneficiaryPage from 'components/Beneficiaries/BeneficiaryPage';
 import { BeneficiaryApplication } from '../../interfaces/interfaces';
+import { IpfsClient } from 'utils/IpfsClient';
 export default function SingleBeneficiaryPage(): JSX.Element {
   const router = useRouter();
   const { contracts } = useContext(ContractsContext);
@@ -13,28 +14,7 @@ export default function SingleBeneficiaryPage(): JSX.Element {
     const ipfsHash = await contracts.beneficiary.getBeneficiary(
       router.query.id as string,
     );
-    const ipfsData = await fetch(
-      `${process.env.IPFS_URL}${getIpfsHashFromBytes32(ipfsHash)}`,
-    ).then((response) => response.json());
-    const beneficiaryApplication: BeneficiaryApplication = {
-      organizationName: ipfsData.name,
-      missionStatement: ipfsData.missionStatement,
-      beneficiaryAddress: ipfsData.beneficiaryAddress,
-      files: {
-        profileImage: ipfsData.profileImage,
-        headerImage: ipfsData?.headerImage,
-        impactReports: ipfsData?.impactReports,
-        additionalImages: ipfsData?.additionalImages,
-      },
-      links: {
-        twitterUrl: ipfsData?.twitterUrl,
-        linkedinUrl: ipfsData?.linkedinUrl,
-        facebookUrl: ipfsData?.linkedinUrl,
-        instagramUrl: ipfsData?.linkedinUrl,
-        githubUrl: ipfsData?.linkedinUrl,
-        proofOfOwnership: ipfsData?.linkedinUrl,
-      },
-    };
+    const beneficiaryApplication = await IpfsClient().get(ipfsHash);
     setBeneficiary(beneficiaryApplication);
   }
