+++ packages/app/pages/beneficiaries/index.tsx
@@ -1,12 +1,14 @@
-import { getIpfsHashFromBytes32 } from '@popcorn/utils/ipfsHashManipulation';
 import BeneficiaryGrid from 'components/Beneficiaries/BeneficiaryGrid';
 import { BeneficiaryApplication } from 'interfaces/interfaces';
 import { useContext, useEffect, useState } from 'react';
+import { IpfsClient } from 'utils/IpfsClient';
 import { ContractsContext } from '../../context/Web3/contracts';
 
 export default function BeneficiaryPage(): JSX.Element {
   const { contracts } = useContext(ContractsContext);
-  const [benefeciaries, setBeneficiaries] = useState<BeneficiaryApplication[]>([]);
+  const [beneficiaries, setBeneficiaries] = useState<BeneficiaryApplication[]>(
+    [],
+  );
 
   async function getBeneficiaries() {
     const beneficiaryAddresses =
@@ -18,32 +20,12 @@ export default function BeneficiaryPage(): JSX.Element {
     );
     const beneficiaryData = await (
       await Promise.all(
-        ipfsHashes.map((ipfsHash) => {
-          return fetch(
-            `${process.env.IPFS_URL}${getIpfsHashFromBytes32(ipfsHash)}`,
-          ).then((response) => response.json());
-        }),
+        ipfsHashes.map(async (cid) => await IpfsClient().get(cid)),
       )
-    ).map((beneficiaryJson) => {
-      return {
-        organizationName: beneficiaryJson.name,
-        missionStatement: beneficiaryJson.missionStatement,
-        beneficiaryAddress: beneficiaryAddresses[0], // TODO: revert to beneficiaryJson.ethereumAddress 
-        files: {
-          profileImage: beneficiaryJson.profileImage,
-          headerImage: beneficiaryJson?.headerImage,
-          impactReports: beneficiaryJson?.impactReports,
-          additionalImages: beneficiaryJson?.additionalImages,
-        },
-        links: {
-          twitterUrl: beneficiaryJson?.twitterUrl,
-          linkedinUrl: beneficiaryJson?.linkedinUrl,
-          facebookUrl: beneficiaryJson?.linkedinUrl,
-          instagramUrl: beneficiaryJson?.linkedinUrl,
-          githubUrl: beneficiaryJson?.linkedinUrl,
-          proofOfOwnership: beneficiaryJson?.linkedinUrl,
-        },
-      }
+    ).map((beneficiaryApplication) => {
+      // TODO: Remove temporary address assignment
+      beneficiaryApplication.beneficiaryAddress = beneficiaryAddresses[0];
+      return beneficiaryApplication;
     });
     setBeneficiaries(beneficiaryData);
   }
@@ -60,7 +42,7 @@ export default function BeneficiaryPage(): JSX.Element {
       subtitle={
         'Beneficiary organizations that have passed the voting process and are eligible to receive grants'
       }
-      beneficiaries={benefeciaries}
+      beneficiaries={beneficiaries}
     />
   );
 }
