+++ packages/app/utils/formatBigNumber.ts
